Rename folder message to messages
