+++ src/ai/llm/messages/formatters/anthropic.ts
