+++ src/ai/llm/messages/formatters/openai.ts
