+++ src/ai/llm/messages/formatters/types.ts
