+++ src/ai/llm/messages/formatters/vercel.ts
